+++ GUI/utils/constants.js
@@ -1,0 +1,15 @@
+export const PADDING = 10;
+export const BORDER_WIDTH = 0.5;
+export const CORNER_RADIUS = 10;
+
+export const CATEGORY_HEIGHT = 30;
+export const CATEGORY_PADDING = 5;
+export const CATEGORY_OFFSET_Y = 50;
+export const CATEGORY_BOX_PADDING = 5;
+
+export const ITEM_SPACING = 5;
+export const SEPARATOR_HEIGHT = 20;
+export const SUBCATEGORY_BUTTON_HEIGHT = 20;
+export const SUBCATEGORY_BUTTON_SPACING = 5;
+
+export const ANIMATION_DURATION = 300;
