+++ Utility/Router.js
@@ -3,7 +3,7 @@ import { Utils } from './Utils';
 import { Chat } from './Chat';
 import { Vec3d } from './Constants';
 import { File } from './Constants';
-import { closeGui } from '../GUI/GUI';
+import { closeGui } from '../GUI/index';
 
 class Routes {
     constructor() {}
